+++ hypervisor/arch/x86/exception.c
@@ -4,6 +4,10 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
+#include <util.h>
+#include <x86/lib/spinlock.h>
+#include <x86/cpu.h>
+#include <irq.h>
 #include <x86/irq.h>
 #include <dump.h>
 #include <logmsg.h>
