+++ hypervisor/arch/x86/guest/ept.c
@@ -7,10 +7,14 @@
 #include <types.h>
 #include <errno.h>
 #include <x86/guest/vm.h>
+#include <util.h>
+#include <x86/lib/spinlock.h>
+#include <x86/cpu.h>
 #include <irq.h>
 #include <x86/pgtable.h>
 #include <x86/mmu.h>
 #include <x86/guest/ept.h>
+#include <x86/guest/virq.h>
 #include <x86/vmx.h>
 #include <x86/vtd.h>
 #include <logmsg.h>
