+++ hypervisor/arch/x86/guest/ucode.c
@@ -6,12 +6,17 @@
 
 #include <types.h>
 #include <errno.h>
+#include <util.h>
 #include <x86/lib/spinlock.h>
 #include <x86/cpu.h>
 #include <x86/msr.h>
 #include <x86/cpuid.h>
 #include <x86/guest/ucode.h>
 #include <x86/guest/guest_memory.h>
+#include <x86/guest/virq.h>
+#include <x86/lib/spinlock.h>
+#include <x86/cpu.h>
+#include <irq.h>
 #include <x86/irq.h>
 #include <logmsg.h>
