+++ hypervisor/arch/x86/guest/vcpu.c
@@ -7,6 +7,7 @@
 #include <types.h>
 #include <errno.h>
 #include <x86/guest/vcpu.h>
+#include <x86/guest/virq.h>
 #include <x86/lib/bits.h>
 #include <x86/vmx.h>
 #include <logmsg.h>
