+++ hypervisor/arch/x86/guest/vlapic.c
@@ -38,6 +38,7 @@
 #include <x86/lapic.h>
 #include <x86/guest/vmcs.h>
 #include <x86/guest/vlapic.h>
+#include <x86/guest/virq.h>
 #include <ptintr.h>
 #include <x86/vmx.h>
 #include <x86/guest/vm.h>
