+++ hypervisor/arch/x86/irq.c
@@ -5,8 +5,13 @@
  */
 
 #include <types.h>
+#include <util.h>
+#include <acrn_common.h>
+#include <x86/lib/spinlock.h>
 #include <x86/per_cpu.h>
 #include <x86/io.h>
+#include <x86/cpu.h>
+#include <irq.h>
 #include <x86/irq.h>
 #include <x86/idt.h>
 #include <x86/ioapic.h>
@@ -15,7 +20,11 @@
 #include <logmsg.h>
 #include <x86/vmx.h>
 
+#define DBG_LEVEL_X86_IRQ 	6U
+
 static spinlock_t x86_irq_spinlock = { .head = 0U, .tail = 0U, };
+
+typedef void (*spurious_handler_t)(uint32_t vector);
 
 spurious_handler_t spurious_handler;
 
@@ -131,7 +140,7 @@ static void free_irq_vector(uint32_t irq)
 void free_irq_arch(uint32_t irq)
 {
 	if (irq < NR_IRQS) {
-		dev_dbg(DBG_LEVEL_IRQ, "[%s] irq%d vr:0x%x",
+		dev_dbg(DBG_LEVEL_X86_IRQ, "[%s] irq%d vr:0x%x",
 			__func__, irq, irq_to_vector(irq));
 		free_irq_vector(irq);
 	}
@@ -280,4 +289,50 @@ void init_irq_descs_arch(struct irq_desc descs[])
 void setup_irqs_arch(void)
 {
 	ioapic_setup_irqs();
+}
+
+static void disable_pic_irqs(void)
+{
+	pio_write8(0xffU, 0xA1U);
+	pio_write8(0xffU, 0x21U);
+}
+
+static inline void fixup_idt(const struct host_idt_descriptor *idtd)
+{
+	uint32_t i;
+	struct idt_64_descriptor *idt_desc = idtd->idt->host_idt_descriptors;
+	uint32_t entry_hi_32, entry_lo_32;
+
+	for (i = 0U; i < HOST_IDT_ENTRIES; i++) {
+		entry_lo_32 = idt_desc[i].offset_63_32;
+		entry_hi_32 = idt_desc[i].rsvd;
+		idt_desc[i].rsvd = 0U;
+		idt_desc[i].offset_63_32 = entry_hi_32;
+		idt_desc[i].high32.bits.offset_31_16 = entry_lo_32 >> 16U;
+		idt_desc[i].low32.bits.offset_15_0 = entry_lo_32 & 0xffffUL;
+	}
+}
+
+static inline void set_idt(struct host_idt_descriptor *idtd)
+{
+	asm volatile ("   lidtq %[idtd]\n" :	/* no output parameters */
+		      :		/* input parameters */
+		      [idtd] "m"(*idtd));
+}
+
+void init_interrupt_arch(uint16_t pcpu_id)
+{
+	struct host_idt_descriptor *idtd = &HOST_IDTR;
+
+	if (pcpu_id == BSP_CPU_ID) {
+		fixup_idt(idtd);
+	}
+	set_idt(idtd);
+	init_lapic(pcpu_id);
+
+	if (pcpu_id == BSP_CPU_ID) {
+		/* we use ioapic only, disable legacy PIC */
+		disable_pic_irqs();
+		init_ioapic();
+	}
 }
