+++ hypervisor/boot/guest/vboot_info.c
@@ -7,7 +7,10 @@
 #include <types.h>
 #include <rtl.h>
 #include <errno.h>
+#include <util.h>
 #include <x86/per_cpu.h>
+#include <x86/lib/spinlock.h>
+#include <x86/cpu.h>
 #include <x86/irq.h>
 #include <multiboot.h>
 #include <x86/pgtable.h>
