+++ hypervisor/include/arch/x86/per_cpu.h
@@ -9,7 +9,12 @@
 
 #include <types.h>
 #include <sbuf.h>
+#include <util.h>
+#include <x86/lib/spinlock.h>
+#include <x86/cpu.h>
+#include <irq.h>
 #include <x86/irq.h>
+#include <x86/notify.h>
 #include <x86/page.h>
 #include <timer.h>
 #include <profiling.h>
